Add an empty integration test step; modify start/stop scripts for greater flexibility

+++ start.sh
@@ -7,9 +7,16 @@ for i in $(seq 1 $N); do
   PROXY=$(( 7000 + i ))
   CONTROL=$(( 3000 + i ))
   mkdir -p data/redis-$REDIS
-  redis-server <(sed 's/{{PORT}}/'$REDIS'/g' redis.conf.in)
+
+  if [[ ! -f data/redis-$REDIS/redis-$REDIS.pid ]]; then
+    redis-server <(sed 's/{{PORT}}/'$REDIS'/g' redis.conf.in)
+  fi
 
   mkdir -p data/hive-$CONTROL
-  ./bin/redis-hive $REDIS $PROXY $CONTROL $(seq 3001 $(( 3000 + N)) ) > data/hive-$CONTROL/hive.log 2>&1 &
-  echo $! > data/hive-$CONTROL/hive.pid
+
+  if [[ ! -f data/hive-$CONTROL/hive.pid ]] ||
+     [[ $(cat /proc/$(cat data/hive-$CONTROL/hive.pid)/comm 2>/dev/null) != redis-hive ]]; then
+    ./bin/redis-hive $REDIS $PROXY $CONTROL $(seq 3001 $(( 3000 + N)) ) > data/hive-$CONTROL/hive.log 2>&1 &
+    echo $! > data/hive-$CONTROL/hive.pid
+  fi
 done
